+++ iroh-gateway/src/handlers.rs
@@ -14,7 +14,7 @@ use handlebars::Handlebars;
 use http::Method;
 use iroh_metrics::{core::MRecorder, gateway::GatewayMetrics, inc, resolver::OutMetrics};
 use iroh_resolver::resolver::{CidOrDomain, UnixfsType};
-use iroh_unixfs::{content_loader::ContentLoader, unixfs::Link};
+use iroh_unixfs::{content_loader::ContentLoader, Link};
 use iroh_util::human::format_bytes;
 use serde::{Deserialize, Serialize};
 use serde_json::{
